refactor: updated fabs in home
